+++ spec/acceptance/api/v2/contacts/merges_spec.rb
@@ -85,6 +85,7 @@ resource 'Contacts > Merges' do
       donor_accounts
       last_six_donations
       people
+      tasks
     )
   end
   # List out any additional attribute keys that will be alongside