+++ spec/acceptance/api/v2/contacts_spec.rb
@@ -111,6 +111,7 @@ resource 'Contacts' do
       donor_accounts
       last_six_donations
       people
+      tasks
     )
   end